+++ src/Resources/app/storefront/src/main.js
@@ -1,5 +1,4 @@
 import ListrakTracking from './plugins/listrak-tracking/listrak-tracking';
-// import ListrakNewsletterCheckbox from './plugins/listrak-newsletter-checkbox/listrak-newsletter-checkbox';
 
 window.PluginManager.register(
     'ListrakTracking',
@@ -15,9 +14,3 @@ window.PluginManager.register(
         ),
     '[data-listrak-newsletter-checkbox]'
 );
-
-// window.PluginManager.register(
-//     'ListrakNewsletterCheckbox',
-//     ListrakNewsletterCheckbox,
-//     '[data-listrak-newsletter-checkbox]'
-// );