+++ src/Resources/app/storefront/src/plugins/listrak-tracking/listrak-tracking.js
@@ -10,10 +10,57 @@ export default class ListrakTracking extends PluginBaseClass {
 
     init() {
         let merchant = this.options.merchantId;
+        let requiresCookieConsent = this.options.requiresCookieConsent;
+        let email = this.options.email;
         const listrakCookie = CookieStorage.getItem(
             this.options.listrakTrackingCookie
         );
-        if (listrakCookie && merchant) {
+        if (merchant) {
+            if (requiresCookieConsent && !listrakCookie) {
+                return;
+            }
+            (function (d, tid, vid) {
+                if (typeof _ltk != 'undefined') {
+                    return;
+                }
+                var js = d.createElement('script');
+                js.id = 'ltkSDK';
+                js.src =
+                    'https://cdn.listrakbi.com/scripts/script.js?m=' +
+                    tid +
+                    '&v=' +
+                    vid;
+                d.querySelector('head').appendChild(js);
+            })(document, merchant, '1');
+
+            (function (d) {
+                if (typeof _ltk == 'undefined') {
+                    if (document.addEventListener)
+                        document.addEventListener(
+                            'ltkAsyncListener',
+                            function () {
+                                _ltk_util.ready(d);
+                            }
+                        );
+                    else {
+                        e = document.documentElement;
+                        e.ltkAsyncProperty = 0;
+                        e.attachEvent('onpropertychange', function (e) {
+                            if (e.propertyName == 'ltkAsyncProperty') {
+                                _ltk_util.ready(d);
+                            }
+                        });
+                    }
+                } else {
+                    _ltk_util.ready(d);
+                }
+            })(function () {
+                if (email) {
+                    _ltk.SCA.Update('email', email);
+                }
+                _ltk.Activity.AddPageBrowse();
+            });
+
             if (this.options.data) {
                 const placement = this.options.data.placement;
                 if (placement === 'order') {